+++ src/mocks/domains/createMockStoryPost.ts
